Supports 'x' and 'y' keys; sorts data
If a user’s data sets the x and y coordinates with keys, this will
support that. Additionally, each function will sort the regression line
points by their x-coordinate, as Highcharts prefers.

I did not touch the _loess() function, as I was a little confused with
it.

Apologies for all the formatting changes. Must be something sublime
text did.