Error when having multiple line charts
Error when having multiple line charts with ON and OFF Regression
I found a problem when adding more then one series to the regression library, I need to have all ON to have regression and cannot be selective regarding which one I would like to have a regression line.
I made a few changes to prevent adding undefined series in the case of regression being OFF and when the last series was undefined the validation seemed to not address this issue, so, I changed it as well (concat). With this I can now have a mix of series with regression ON and OFF without any problem.